Fixing of naming Mysql (Here is the error, SQL Error: 1054, SQLState: 42S22, Unknown column)feilds from phoneNumber to phone_number by adding annotation @Colum(name="phone_number")

+++ src/main/java/com/yapollo/acountingofservice/repository/ReceiptsRepository.java
@@ -3,9 +3,9 @@ package com.yapollo.acountingofservice.repository;
 import com.yapollo.acountingofservice.model.RepairReceipt;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ReceiptsRepository extends CrudRepository<RepairReceipt, Long> {
+public interface ReceiptsRepository extends CrudRepository<RepairReceipt, Integer> {
 
-    RepairReceipt findByPhoneNumber (int phoneNumber);
+    RepairReceipt findByPhoneNumber (Integer phoneNumber);
 
 
 }
